TypeScript: highlight `interface` in pure TypeScript files
It is getting cumbersome to having to define the same rules separately
to JS files, TS files and their React variants. I'm considering
dropping the file extensions from the definitions, but I'm afraid this
might cause the styles to spill to other file types and I'd detect
this much later when fixing it is hassle.

I'll have to think on this some more.